+++ src/MOCK_DB_DATA/library-css/css-components.data.ts
@@ -1,10 +1,8 @@
 import { LibraryItem } from 'src/app/views/shared/library/models/library-item.model';
 
-import { Toast } from './components/toast';
 import { Tooltip } from './components/tooltip';
 
 export const CssComponentsData: LibraryItem[] = [
-  ...Toast,
   ...Tooltip,
 ];
