+++ src/MOCK_DB_DATA/library-items/toast/_toast.data.ts
@@ -1,0 +1,9 @@
+import { LibraryItem } from 'src/app/views/shared/library/models/library-item.model';
+
+import { Toast } from './toast';
+
+export const ToastData: LibraryItem[] = [
+  ...Toast
+];
+
+
