+++ src/MOCK_DB_DATA/library-items/toast/toast.ts
@@ -1,0 +1,51 @@
+import { LibraryItem, LibraryItemType } from 'src/app/views/shared/library/models/library-item.model';
+import { BootstrapComponentsBaseUrl, NgBootstrapComponentsBaseUrl } from '../../bootstrap-base-urls';
+
+export const Toast: LibraryItem[] = [{
+  id: 'toast',
+  title: 'Toast',
+  type: LibraryItemType.ngBootstrap,
+  exampleHtml: getExampleHtml(),
+  codeHtml: getCodeHtml(),
+  documentationHtml: getDocumentationHtml()
+}];
+
+
+/*
+ * Return value is ignored if LibraryItemType is not html
+ */
+function getExampleHtml(): string {
+  return ``;
+}
+
+/*
+ * Return empty string to use a copy of exampleHtml as codeHtml.
+ * Return null to remove Code from library-item.
+ */
+function getCodeHtml(): string | null {
+  return `
+<div class="toast-container">
+  <ngb-toast *ngIf="show" header="Toast title" [autohide]="true" (hidden)="show=false">
+    Standard toast. This instance is automatically hidden.
+  </ngb-toast>
+</div>`;
+}
+
+/*
+ * Return null to remove Code from library-item.
+ */
+function getDocumentationHtml(): string | null {
+  return `
+<p>
+  Bootstrap documentation for
+  <a href="${BootstrapComponentsBaseUrl}/toasts">Toast</a>
+</p>
+<p>
+  The toast is implemented as a
+  <a href="${NgBootstrapComponentsBaseUrl}/toast">NgBootstrap Toast</a>
+  in the FHI Designsystem.
+</p>
+<p>
+  If you use another Javascript framework, you will have to find a suitable 3. party library.
+</p>`;
+}
