+++ src/app/core/mock-db.service.ts
@@ -8,6 +8,7 @@ import { FormsData } from 'src/MOCK_DB_DATA/library-items/forms/_forms.data';
 import { IconsData } from 'src/MOCK_DB_DATA/library-items/icons/_icons.data';
 import { PaginationData } from 'src/MOCK_DB_DATA/library-items/pagination/_pagination.data';
 import { TableData } from 'src/MOCK_DB_DATA/library-items/table/_table.data';
+import { ToastData } from 'src/MOCK_DB_DATA/library-items/toast/_toast.data';
 import { TypographyData } from 'src/MOCK_DB_DATA/library-items/typography/_typography.data';
 
 import { CssComponentsData } from 'src/MOCK_DB_DATA/library-css/css-components.data';
@@ -26,6 +27,7 @@ export class MockDbService implements InMemoryDbService {
       IconsData,
       PaginationData,
       TableData,
+      ToastData,
       TypographyData,
 
       CssComponentsData,
